requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.0
cachetools==5.5.2
celery==5.3.4
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.40.0
crispy-bootstrap5==2025.6
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.2.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.4
django-filter==24.3
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.3
drf-spectacular==0.27.2
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.30.0
idna==3.7
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.3.4
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
setuptools==80.9.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==10.8.0
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
